libmrc: crds - API CHANGE: in crds{_gen} xl an xh are now double
This is a small change that has a big ripple effect. Pretty much everything
in this library has been upgraded to work, but I expect this will break
in external codes.

+++ include/mrc_crds.h
@@ -8,8 +8,8 @@
 struct mrc_crds {
   struct mrc_obj obj;
   // parameters
-  float xl[3];
-  float xh[3];
+  double xl[3];
+  double xh[3];
   int sw;
 
   // state
